+++ test/core/TestingFunctionMacro_UnitTests.cmake
@@ -816,6 +816,36 @@ function(unittest_tribits_add_test_basic)
     "NAME;${PACKEXEN};COMMAND;${PACKEXEN_PATH};--arg1=\"'bob' and 'cats'\""
     )
 
+  message("Add one test using POSTFIX_AND_ARGS_0 with empty postfix")
+  global_set(TRIBITS_ADD_TEST_ADD_TEST_INPUT)
+  tribits_add_test( ${EXEN}
+    POSTFIX_AND_ARGS_0  ""  arg1
+    ADDED_TESTS_NAMES_OUT  ${EXEN}_TEST_NAMES
+    )
+  unittest_compare_const(
+    TRIBITS_ADD_TEST_ADD_TEST_INPUT
+    "NAME;${PACKEXEN};COMMAND;${PACKEXEN_PATH};arg1"
+    )
+  unittest_compare_const( ${EXEN}_TEST_NAMES  "${PACKEXEN}" )
+
+  message("Add two tests using POSTFIX_AND_ARGS_0 with spaces and semi-colons")
+  global_set(TRIBITS_ADD_TEST_ADD_TEST_INPUT)
+  tribits_add_test( ${EXEN}
+    POSTFIX_AND_ARGS_0  "test0"  "has some<sep>spaces" "--and semi<sep>colons too"
+    POSTFIX_AND_ARGS_1  "test1"  "has2 some<sep>spaces" "--and2 semi<sep>colons too"
+    LIST_SEPARATOR "<sep>"
+    ADDED_TESTS_NAMES_OUT  ${EXEN}_TEST_NAMES
+    )
+  unittest_compare_const(
+    TRIBITS_ADD_TEST_ADD_TEST_INPUT
+    "NAME;PackageA_SomeExec_test0;COMMAND;${PACKEXEN_PATH};has some;spaces;--and semi;colons too;NAME;PackageA_SomeExec_test1;COMMAND;${PACKEXEN_PATH};has2 some;spaces;--and2 semi;colons too"
+    )
+  unittest_compare_const( ${EXEN}_TEST_NAMES  "${PACKEXEN}_test0;${PACKEXEN}_test1" )
+  # NOTE: There is a test under CTestScriptsUnitTests/ that actaully runs an
+  # real test and command that verifies the correct handling of semi-colons.
+  # The mashing that CMake above is doing to these replaced semicolons does
+  # not do them justice.
+
   message("Add two tests with different postfixes and arguments")
   global_set(TRIBITS_ADD_TEST_ADD_TEST_INPUT)
   tribits_add_test( ${EXEN}
@@ -1406,6 +1436,27 @@ function(unittest_tribits_add_test_disable)
   unittest_compare_const(
     TRIBITS_ADD_ADVANCED_TEST_CMND_ARRAY_0
     "\"someCmnd\""
+    )
+  unittest_compare_const(
+    TRIBITS_ADD_ADVANCED_TEST_NUM_CMNDS
+    1
+    )
+
+  message("Check that tribits_add_advanced_test(...) produces FATAL_ERROR with unparsed args")
+  tribits_add_advanced_test_unittest_reset()
+  set(${PROJECT_NAME}_CHECK_FOR_UNPARSED_ARGUMENTS CUSTOM_FAILURE_MODE)
+  tribits_add_advanced_test( SomeCmnd
+     "unparsedarg1" "unparsedarg2"
+    TEST_0 "unparsedarg3" "unparsedarg4" CMND someCmnd ARGS "arg1<semicol>arg2"
+    ADDED_TEST_NAME_OUT  SomeCmnd_TEST_NAME
+    )
+  unittest_compare_const(
+    MESSAGE_WRAPPER_INPUT
+    "CUSTOM_FAILURE_MODE;Arguments passed in unrecognized.  PARSE_UNPARSED_ARGUMENTS = 'unparsedarg1;unparsedarg2';CUSTOM_FAILURE_MODE;Arguments passed in unrecognized.  PARSE_UNPARSED_ARGUMENTS = 'unparsedarg3;unparsedarg4';-- PackageA_SomeCmnd: Added test (BASIC, PROCESSORS=1)!"
+    )
+  unittest_compare_const(
+    TRIBITS_ADD_ADVANCED_TEST_CMND_ARRAY_0
+    "\"someCmnd\" \"arg1<semicol>arg2\""
     )
   unittest_compare_const(
     TRIBITS_ADD_ADVANCED_TEST_NUM_CMNDS
@@ -2388,6 +2439,27 @@ function(unittest_tribits_add_advanced_test_basic)
     "${CMAKE_CURRENT_BINARY_DIR}/${PACKAGE_NAME}_TAAT_basic_exec_1_args_3.cmake"
     REGEX_STRINGS
       "TEST_0_CMND \"${CMAKE_CURRENT_BINARY_DIR}/${PACKEXEN}\" \"arg1\" \"arg2\" \"arg3\""
+      "NUM_CMNDS 1"
+    )
+
+  message("***\n*** Add a single package executable with quoted arguments containing semi-colons\n***")
+  tribits_add_advanced_test_unittest_reset()
+  tribits_add_advanced_test( TAAT_basic_exec_1_args_3_quotes_semicolons
+    TEST_0 EXEC ${EXEN} ARGS "arg1=val1<semicolon>val2" arg2 arg3
+    LIST_SEPARATOR "<semicolon>"
+    )
+  unittest_compare_const(
+    TRIBITS_ADD_ADVANCED_TEST_CMND_ARRAY_0
+    "\"${CMAKE_CURRENT_BINARY_DIR}/${PACKEXEN}\" \"arg1=val1<semicolon>val2\" \"arg2\" \"arg3\""
+    )
+  unittest_compare_const(
+    TRIBITS_ADD_ADVANCED_TEST_NUM_CMNDS
+    1
+    )
+  unittest_file_regex(
+    "${CMAKE_CURRENT_BINARY_DIR}/${PACKAGE_NAME}_TAAT_basic_exec_1_args_3_quotes_semicolons.cmake"
+    REGEX_STRINGS
+      "TEST_0_CMND \"${CMAKE_CURRENT_BINARY_DIR}/${PACKEXEN}\" \"arg1=val1<semicolon>val2\" \"arg2\" \"arg3\""
       "NUM_CMNDS 1"
     )
 
@@ -3938,37 +4010,38 @@ function(unittest_tribits_add_executable_and_test)
 
   tribits_add_executable_and_test(
     execName
-    SOURCES src1 src2
-    NAME testName
-    NAME_POSTFIX testNamePostfix
-    CATEGORIES category1 category2
-    HOST host1 host2
-    DEPLIBS lib1 lib2  # Deprecated
-    XHOST host1 host2
-    HOSTTYPE hosttype1 hosttype2
-    XHOSTTYPE hosttype1 hosttype2
-    EXCLUDE_IF_NOT_TRUE var1 var2
-    DISABLED "Disable this test because I said"
-    NOEXEPREFIX
-    NOEXESUFFIX
-    IMPORTEDLIBS ilib1 ilib2
-    TESTONLYLIBS tolib1 tolib2
-    DIRECTORY dir
-    COMM serial mpi
-    NUM_MPI_PROCS numProcs
-    LINKER_LANGUAGE C
-    ADD_DIR_TO_NAME
-    TARGET_DEFINES -DSOMEDEFINE1
-    DEFINES -DSOMEDEFINE2
-    KEYWORDS keyword1 keyword2
-    STANDARD_PASS_OUTPUT
-    PASS_REGULAR_EXPRESSION "regex1;regex2"
-    FAIL_REGULAR_EXPRESSION "regex1;regex2"
-    ENVIRONMENT env1=envval1 env2=envval2
-    WILL_FAIL
-    TIMEOUT 11.5
-    ADDED_EXE_TARGET_NAME_OUT execName_TARGET_NAME
     ADDED_TESTS_NAMES_OUT execName_TEST_NAME
+    ADDED_EXE_TARGET_NAME_OUT execName_TARGET_NAME
+    LIST_SEPARATOR <semicolon>
+    TIMEOUT 11.5
+    WILL_FAIL
+    ENVIRONMENT env1=envval1 env2=envval2
+    FAIL_REGULAR_EXPRESSION "regex1;regex2"
+    PASS_REGULAR_EXPRESSION "regex1;regex2"
+    STANDARD_PASS_OUTPUT
+    KEYWORDS keyword1 keyword2
+    DEFINES -DSOMEDEFINE2
+    TARGET_DEFINES -DSOMEDEFINE1
+    ADD_DIR_TO_NAME
+    LINKER_LANGUAGE C
+    NUM_MPI_PROCS numProcs
+    COMM serial mpi
+    DIRECTORY dir
+    TESTONLYLIBS tolib1 tolib2
+    IMPORTEDLIBS ilib1 ilib2
+    NOEXESUFFIX
+    NOEXEPREFIX
+    DISABLED "Disable this test because I said"
+    EXCLUDE_IF_NOT_TRUE var1 var2
+    XHOSTTYPE hosttype1 hosttype2
+    HOSTTYPE hosttype1 hosttype2
+    XHOST host1 host2
+    DEPLIBS lib1 lib2  # Deprecated
+    HOST host1 host2
+    CATEGORIES category1 category2
+    NAME_POSTFIX testNamePostfix
+    NAME testName
+    SOURCES src1 src2
     )
   unittest_compare_const(
     TRIBITS_ADD_EXECUTABLE_CAPTURE_ARGS
@@ -3976,8 +4049,10 @@ function(unittest_tribits_add_executable_and_test)
     )
   unittest_compare_const(
     TRIBITS_ADD_TEST_CAPTURE_ARGS
-    "execName;COMM;serial;mpi;CATEGORIES;category1;category2;HOST;host1;host2;XHOST;host1;host2;HOSTTYPE;hosttype1;hosttype2;XHOSTTYPE;hosttype1;hosttype2;EXCLUDE_IF_NOT_TRUE;var1;var2;NOEXEPREFIX;NOEXESUFFIX;NAME;testName;NAME_POSTFIX;testNamePostfix;DIRECTORY;dir;KEYWORDS;keyword1;keyword2;NUM_MPI_PROCS;numProcs;PASS_REGULAR_EXPRESSION;regex1;regex2;FAIL_REGULAR_EXPRESSION;regex1;regex2;ENVIRONMENT;env1=envval1;env2=envval2;DISABLED;Disable this test because I said;STANDARD_PASS_OUTPUT;WILL_FAIL;TIMEOUT;11.5;ADD_DIR_TO_NAME;ADDED_TESTS_NAMES_OUT;ADDED_TESTS_NAMES_OUT;ADDED_TESTS_NAMES"
+    "execName;COMM;serial;mpi;CATEGORIES;category1;category2;HOST;host1;host2;XHOST;host1;host2;HOSTTYPE;hosttype1;hosttype2;XHOSTTYPE;hosttype1;hosttype2;EXCLUDE_IF_NOT_TRUE;var1;var2;NOEXEPREFIX;NOEXESUFFIX;NAME;testName;NAME_POSTFIX;testNamePostfix;DIRECTORY;dir;KEYWORDS;keyword1;keyword2;NUM_MPI_PROCS;numProcs;PASS_REGULAR_EXPRESSION;regex1;regex2;FAIL_REGULAR_EXPRESSION;regex1;regex2;ENVIRONMENT;env1=envval1;env2=envval2;DISABLED;Disable this test because I said;STANDARD_PASS_OUTPUT;WILL_FAIL;TIMEOUT;11.5;LIST_SEPARATOR;<semicolon>;ADD_DIR_TO_NAME;ADDED_TESTS_NAMES_OUT;ADDED_TESTS_NAMES_OUT;ADDED_TESTS_NAMES"
     )
+  # NOTE: Above, we input the list in reverse order to prove that the
+  # arguments are handled correctly internally.
 
   message("\n***")
   message("*** Test passing in XHOST_TEST and XHOSTTYPE_TEST into tribits_add_executable_and_test(...)")
@@ -4438,4 +4513,4 @@ message("*** Determine final result of all unit tests")
 message("***\n")
 
 # Pass in the number of expected tests that must pass!
-unittest_final_result(675)
+unittest_final_result(686)